# docs/swagger.yaml
definitions:
  models.Token:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
...
  request.CreateUser:
    properties:
      email:
        type: string
      first_name:
        maxLength: 20
        minLength: 1
        type: string
      last_name:
        maxLength: 20
        minLength: 1
        type: string
      password:
        maxLength: 20
        minLength: 8
        type: string
      user_role_id:
        type: integer
      username:
        maxLength: 20
        minLength: 4
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    - user_role_id
    - username
    type: object
...
  request.RefreshToken:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  request.SignIn:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
  request.UpdateUser:
    properties:
      email:
        type: string
      first_name:
        maxLength: 20
        minLength: 1
        type: string
      is_active:
        type: boolean
      last_name:
        maxLength: 20
        minLength: 1
        type: string
      password:
        maxLength: 20
        minLength: 8
        type: string
      user_role_id:
        type: integer
      username:
        maxLength: 20
        minLength: 4
        type: string
    type: object
...
  response.GetUser:
    properties:
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      last_name:
        type: string
      user_role:
        type: integer
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  title: Canteen Management System API
  version: "1.0"
paths:
  /api/auth/refresh-token:
    post:
      consumes:
      - application/json
      description: Refresh access and refresh tokens with the provided refresh token
      parameters:
      - description: Refresh token object
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.RefreshToken'
      produces:
      - application/json
      responses:
        "200":
          description: Successful response
          schema:
            $ref: '#/definitions/models.Token'
        "400":
          description: Invalid input JSON
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Refresh access and refresh tokens
      tags:
      - auth
  /api/auth/sign-in:
    post:
      consumes:
      - application/json
      description: Sign in a user with the provided JSON input
      parameters:
      - description: User sign in object
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.SignIn'
      produces:
      - application/json
      responses:
        "200":
          description: Successful response
          schema:
            $ref: '#/definitions/models.Token'
        "400":
          description: Invalid input JSON
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Sign in a user
      tags:
      - auth
  /api/auth/sign-out:
    post:
      consumes:
      - application/json
      description: Sign out a user with the provided refresh token
      parameters:
      - description: Refresh token object
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.RefreshToken'
      produces:
      - application/json
      responses:
        "200":
          description: User signed out
          schema:
            type: string
        "400":
          description: Invalid input JSON
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Sign out a user
      tags:
      - auth
...
  /api/users:
    get:
      consumes:
      - application/json
      description: Get all users available
      operationId: get-all-users
      produces:
      - application/json
      responses:
        "200":
          description: Successful response
          schema:
            items:
              $ref: '#/definitions/response.GetUser'
            type: array
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user with the provided JSON input
      parameters:
      - description: User object to be created
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.CreateUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create a new user
      tags:
      - users
  /api/users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user based on ID
      operationId: delete-user
      parameters:
      - description: User ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete a user by ID
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Get a user based on ID
      operationId: get-user-by-id
      parameters:
      - description: User ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successful response
          schema:
            $ref: '#/definitions/response.GetUser'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get a user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update the existing user with the provided JSON input
      parameters:
      - description: User ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: User object to be updated
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/request.UpdateUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Update the existing user
      tags:
      - users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"